refactor workflow tests
move nested to their own test class
