+++ core/src/test/java/org/apache/brooklyn/core/workflow/WorkflowInputOutputExtensionTest.java
@@ -41,10 +41,8 @@ import org.apache.brooklyn.util.core.config.ConfigBag;
 import org.apache.brooklyn.util.exceptions.Exceptions;
 import org.apache.brooklyn.util.text.StringEscapes;
 import org.apache.brooklyn.util.text.Strings;
-import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.function.BiFunction;
@@ -142,101 +140,6 @@ public class WorkflowInputOutputExtensionTest extends BrooklynMgmtUnitTestSuppor
             Task<?> invocation = app.invoke(app.getEntityType().getEffectorByName("myWorkflow").get(), null);
             return invocation.getUnchecked();
         }
-    }
-
-    void assertLogStepMessages(String ...lines) {
-        Assert.assertEquals(lastLogWatcher.getMessages(),
-                Arrays.asList(lines));
-    }
-
-    @Test
-    public void testNestedWorkflowBasic() throws Exception {
-        Object output = invokeWorkflowStepsWithLogging(MutableList.of(
-                MutableMap.of("type", "workflow",
-                        "steps", MutableList.of("return done"))));
-        Asserts.assertEquals(output, "done");
-    }
-
-    @Test
-    public void testNestedWorkflowParametersForbiddenWhenUsedDirectly() throws Exception {
-        Asserts.assertFailsWith(() -> invokeWorkflowStepsWithLogging(MutableList.of(
-                MutableMap.of("type", "workflow",
-                        "parameters", MutableMap.of(),
-                        "steps", MutableList.of("return done")))),
-                e -> Asserts.expectedFailureContainsIgnoreCase(e, "parameters"));
-    }
-
-    @Test
-    public void testExtendingAStepWhichWorksButIsMessyAroundParameters() throws Exception {
-        /*
-         * extending any step type is supported, but discouraged because of confusion and no parameter definitions;
-         * the preferred way is to use the method in the following test
-         */
-        addBeanWithType("log-hi", "1", Strings.lines(
-                "type: log",
-                "message: hi ${name}",
-                "input:",
-                "  name: you"
-        ));
-
-        invokeWorkflowStepsWithLogging(MutableList.of(MutableMap.of("type", "log-hi", "input", MutableMap.of("name", "bob"))));
-        assertLogStepMessages("hi bob");
-
-        invokeWorkflowStepsWithLogging(MutableList.of(MutableMap.of("type", "log-hi")));
-        assertLogStepMessages("hi you");
-    }
-
-    @Test
-    public void testDefiningCustomWorkflowStep() throws Exception {
-        addBeanWithType("log-hi", "1", Strings.lines(
-                "type: workflow",
-                "parameters:",
-                "  name: {}",
-                "steps:",
-                "  - log hi ${name}"
-        ));
-        invokeWorkflowStepsWithLogging(MutableList.of(MutableMap.of("type", "log-hi", "input", MutableMap.of("name", "bob"))));
-        assertLogStepMessages("hi bob");
-
-        Asserts.assertFailsWith(() -> invokeWorkflowStepsWithLogging(MutableList.of(
-                        MutableMap.of("type", "log-hi",
-                                "steps", MutableList.of("return not allowed to override")))),
-                e -> Asserts.expectedFailureContainsIgnoreCase(e, "steps"));
-    }
-
-    @Test
-    public void testDefiningCustomWorkflowStepWithShorthand() throws Exception {
-        addBeanWithType("log-hi", "1", Strings.lines(
-                "type: workflow",
-                "shorthand: ${name}",
-                "parameters:",
-                "  name: {}",
-                "steps:",
-                "  - log hi ${name}"
-        ));
-        invokeWorkflowStepsWithLogging(MutableList.of("log-hi bob"));
-        assertLogStepMessages("hi bob");
-    }
-
-    @Test
-    public void testDefiningCustomWorkflowStepWithOutput() throws Exception {
-        addBeanWithType("log-hi", "1", Strings.lines(
-                "type: workflow",
-                "parameters:",
-                "  name: {}",
-                "steps:",
-                "  - log hi ${name}",
-                "output:",
-                "  message: hi ${name}"
-        ));
-        Object output = invokeWorkflowStepsWithLogging(MutableList.of(MutableMap.of("type", "log-hi", "input", MutableMap.of("name", "bob"))));
-        assertLogStepMessages("hi bob");
-        Asserts.assertEquals(output, MutableMap.of("message", "hi bob"));
-
-        // output can be overridden
-        output = invokeWorkflowStepsWithLogging(MutableList.of(MutableMap.of("type", "log-hi", "input", MutableMap.of("name", "bob"), "output", "I said ${message}")));
-        assertLogStepMessages("hi bob");
-        Asserts.assertEquals(output, "I said hi bob");
     }
 
     // test complex object in an expression
